fix interoperability with Go 1.15 encoding/asn1 SET OF ordering
X690 Section 11.6 specifies a required order for SET OFs in DER.
Go 1.15 updated encoding/asn1 to perform this ordering, which caused
a discrepancy between the order of the attributes in SignerInfo
and the order in the marshaled form generated by MarshaledForSigning().

The attributes are now sorted before calling MarshaledforSigning()
to ensure consistency. This fixes interoperability with other systems
(e.g. OpenSSL CMS) for signatures generated by ietf-cms when compiled
with Go >= 1.15.

In order to verify signatures created by Go < 1.15 on Go >= 1.15,
we need to preserve the received order of SignedAttributes. This
adds a MarshaledForVerification() method to Attributes that behaves
similarly to OpenSSL's CMS_Attributes_Verify template in cms_asn1.c.

This change also adds tests for interoperability with Go < 1.15,
Go >= 1.15, and OpenSSL CMS.

Fixes #25